Update default.frag with fixed comment
There is a comment that has inaccurate information - it claims that the y-coordinate of the mouse starts at the top at 0 pixels and increases as it goes down, however the y-coordinate of the mouse starts at the bottom at 0 pixels and increases as it goes up. This change is to prevent new users learning glsl from being confused by the behavior of the mouse.